feat: :sparkles: added permission controller functions

+++ src/rbac/permission.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { RoleBasedAccessControlRepository } from './rbac.repository';
 import {
   AssignPermissionToRoleDto,
@@ -11,6 +11,14 @@ import {
 export class PermissionService {
   constructor(private readonly rbacRepository: RoleBasedAccessControlRepository) {}
 
+  async checkIfPermissionExists(permissionId: string) {
+    const permission = await this.rbacRepository.getPermissionDetails(permissionId);
+    if (!permission) {
+      throw new NotFoundException('Permission with provided details not found.');
+    }
+    return permission;
+  }
+
   async createPermission({ name, description }: CreatePermissionDto) {
     const permission = await this.rbacRepository.createPermission({
       name,
@@ -20,22 +28,26 @@ export class PermissionService {
   }
 
   async assignPermissionToRole({ permissionId, roleId }: AssignPermissionToRoleDto) {
+    await this.checkIfPermissionExists(permissionId);
     const permission = await this.rbacRepository.assignPermissionToARole(permissionId, roleId);
     return { success: true, permission };
   }
 
-  async getPermission({ permissionId }: PermissionIdDo) {
-    const permission = await this.rbacRepository.getSpecificPermission(permissionId);
-    return permission;
+  async getPermissionDetails({ permissionId }: PermissionIdDo) {
+    await this.checkIfPermissionExists(permissionId);
+    const permission = await this.rbacRepository.getPermissionDetails(permissionId);
+    return { success: true, permission };
   }
 
   async updatePermission({ permissionId, ...data }: UpdatePermissionDto) {
+    await this.checkIfPermissionExists(permissionId);
     const permission = await this.rbacRepository.updatePermission(permissionId, data);
-    return permission;
+    return { success: true, permission };
   }
 
   async deletePermission({ permissionId }: PermissionIdDo) {
+    await this.checkIfPermissionExists(permissionId);
     const permission = await this.rbacRepository.deletePermission(permissionId);
-    return permission;
+    return { success: true, permission };
   }
 }
